UserHome.php
UserHome.php displays a table for pending requests for the currently logged on user and a table for completed requests for the currenlty logged on user. Added the cancel button where a user can cancel a request which deletes the matching row from the orders table in the database. Can view the requests in invoice form. Assuming userID=104 until I can pull the ID from the session.